Show inlayhints for ATOM nodes
